Putting Texnical's change back
This should ease up maintenance and readability of the code, so that you're focusing only on important bits of the code. The code is essentially broken up into:

    Helper classes
    Relevant data
    Class properties
    Helper methods
    Handler subroutine methods
    Event handlers

This "relevant data" portion could be parsed out into a database, whether it's SQLite, MySQL, or otherwise. This means you could update the data without touching the codebase.